adding gizmos
